+++ python/cudf/cudf/bindings/unaryops.pyx
@@ -18,6 +18,11 @@ from libcpp.string cimport string
 
 import numpy as np
 
+from cudf.utils import cudautils
+import numba.cuda
+import numba.numpy_support
+
+
 _UNARY_OP = {
     'sin': SIN,
     'cos': COS,
@@ -46,35 +51,34 @@ def apply_unary_op(incol, op):
     cdef gdf_column* c_incol = column_view_from_column(incol)
 
     cdef gdf_column result
-    cdef unary_op c_op = _UNARY_OP[op]
-    with nogil:
-        result = unary_operation(
-            c_incol[0],
-            c_op
-        )
+
+    cdef unary_op c_op
+    cdef string cpp_str
+    cdef gdf_dtype g_type
+
+    if callable(op):
+        nb_type = numba.numpy_support.from_dtype(incol.dtype)
+        type_signature = (nb_type,)
+        compiled_op = cudautils.compile_udf(op, type_signature)
+        cpp_str = compiled_op[0].encode('UTF-8')
+        if compiled_op[1] not in dtypes:
+            raise TypeError(
+                "Result of window function has unsupported dtype {}"
+                .format(op[1])
+            )
+        g_type = dtypes[compiled_op[1]]
+        with nogil:
+            result = transform(c_incol[0], cpp_str, g_type, True)
+    else:
+        c_op = _UNARY_OP[op]
+        with nogil:
+            result = unary_operation(
+                c_incol[0],
+                c_op
+            )
 
     free(c_incol)
     data, mask = gdf_column_to_column_mem(&result)
-    return Column.from_mem_views(data, mask)
-
-
-def column_applymap(incol, udf_ptx, np_dtype):
-
-    cdef gdf_column* c_incol = column_view_from_column(incol)
-
-    cdef string cpp_str = udf_ptx.encode('UTF-8')
-    cdef gdf_column c_outcol
-
-    # get the gdf_type related to the input np type
-    cdef gdf_dtype g_type = dtypes[np_dtype]
-
-    with nogil:
-        c_outcol = transform(<gdf_column>c_incol[0], cpp_str, g_type, True)
-
-    data, mask = gdf_column_to_column_mem(&c_outcol)
-
-    free(c_incol)
-
     return Column.from_mem_views(data, mask)