+++ python/cudf/cudf/tests/test_cuda_array_interface.py
@@ -26,20 +26,13 @@ basic_dtypes = [
     np.dtype("float32"),
     np.dtype("float64"),
 ]
-string_dtypes = [np.dtype("object")]
+string_dtypes = [np.dtype("str")]
 datetime_dtypes = [
-    # np.dtype("datetime64[ns]"),
-    # np.dtype("datetime64[us]"),
+    np.dtype("datetime64[ns]"),
+    np.dtype("datetime64[us]"),
     np.dtype("datetime64[ms]"),
-    # np.dtype("datetime64[s]"),
+    np.dtype("datetime64[s]"),
 ]
-
-
-def data_type_expectation_builder(data):
-    if data.size > 0:
-        return does_not_raise()
-    else:
-        return pytest.raises((ValueError, IndexError))
 
 
 @pytest.mark.parametrize("dtype", basic_dtypes + datetime_dtypes)
@@ -51,24 +44,25 @@ def test_cuda_array_interface_interop_in(dtype, module):
     if module == "cupy":
         if not _have_cupy:
             pytest.skip("no cupy")
+        module_constructor = cupy.array
         if dtype in datetime_dtypes:
-            expectation = pytest.raises([KeyError])
-        with expectation:
-            module_data = cupy.array(np_data)
+            expectation = pytest.raises(ValueError)
     elif module == "numba":
-        module_data = cuda.to_device(np_data)
+        module_constructor = cuda.to_device
 
-    pd_data = pd.Series(np_data)
+    with expectation:
+        module_data = module_constructor(np_data)
 
-    # Test using a specific function for __cuda_array_interface__ here
-    cudf_data = cudf.Series(module_data)
+        pd_data = pd.Series(np_data)
+        # Test using a specific function for __cuda_array_interface__ here
+        cudf_data = cudf.Series(module_data)
 
-    assert_eq(pd_data, cudf_data)
+        assert_eq(pd_data, cudf_data)
 
-    gdf = cudf.DataFrame()
-    gdf["test"] = module_data
-    pd_data.name = "test"
-    assert_eq(pd_data, gdf["test"])
+        gdf = cudf.DataFrame()
+        gdf["test"] = module_data
+        pd_data.name = "test"
+        assert_eq(pd_data, gdf["test"])
 
 
 @pytest.mark.parametrize(
@@ -76,27 +70,67 @@ def test_cuda_array_interface_interop_in(dtype, module):
 )
 @pytest.mark.parametrize("module", ["cupy", "numba"])
 def test_cuda_array_interface_interop_out(dtype, module):
-    np_data = np.arange(10).astype(dtype)
-    cudf_data = cudf.Series(np_data)
-
     expectation = does_not_raise()
     if dtype in string_dtypes:
         expectation = pytest.raises(NotImplementedError)
     if module == "cupy":
         if not _have_cupy:
             pytest.skip("no cupy")
-        if dtype in datetime_dtypes:
-            expectation = pytest.raises([KeyError])
-        with expectation:
-            module_data = cupy.asarray(cudf_data)
-            got = cupy.asnumpy(module_data)
-    elif module == "numba":
-        with expectation:
-            module_data = cuda.as_cuda_array(cudf_data)
-            got = module_data.copy_to_host()
+        module_constructor = cupy.asarray
 
-    expect = np_data
-    assert_eq(expect, got)
+        def to_host_function(x):
+            return cupy.asnumpy(x)
+
+    elif module == "numba":
+        module_constructor = cuda.as_cuda_array
+
+        def to_host_function(x):
+            return x.copy_to_host()
+
+    with expectation:
+        np_data = np.arange(10).astype(dtype)
+        cudf_data = cudf.Series(np_data)
+        assert isinstance(cudf_data.__cuda_array_interface__, dict)
+
+        module_data = module_constructor(cudf_data)
+        got = to_host_function(module_data)
+
+        expect = np_data
+
+        assert_eq(expect, got)
+
+
+@pytest.mark.parametrize("dtype", basic_dtypes + datetime_dtypes)
+@pytest.mark.parametrize("module", ["cupy", "numba"])
+def test_cuda_array_interface_interop_out_masked(dtype, module):
+    expectation = does_not_raise()
+    if module == "cupy":
+        pytest.skip(
+            "cupy doesn't support version 1 of "
+            "`__cuda_array_interface__` yet"
+        )
+        if not _have_cupy:
+            pytest.skip("no cupy")
+        module_constructor = cupy.asarray
+
+        def to_host_function(x):
+            return cupy.asnumpy(x)
+
+    elif module == "numba":
+        expectation = pytest.raises(NotImplementedError)
+        module_constructor = cuda.as_cuda_array
+
+        def to_host_function(x):
+            return x.copy_to_host()
+
+    np_data = np.arange(10).astype("float64")
+    np_data[[0, 2, 4, 6, 8]] = np.nan
+
+    with expectation:
+        cudf_data = cudf.Series(np_data).astype(dtype)
+        assert isinstance(cudf_data.__cuda_array_interface__, dict)
+
+        module_data = module_constructor(cudf_data)  # noqa: F841
 
 
 @pytest.mark.parametrize("dtype", basic_dtypes + datetime_dtypes)